Add files via upload
Replay files not included. Requires node as well as the listed slippi JS package installed to go along with the conda environment.

This is a continuation of a previously worked on project, mostly done as a proof of concept. The script creates a local SQL database in memory using duckDB (although it could also make a permanent one through a DB file if desired), and writes metadata of each replay file to a table, which then is outputted as a CSV, although if you go the route of the DB file this can be skipped. Originally, this was done without a local SQL database and done purely through parsing and Pandas, but that turned out to be a very time and memory intensive method.